requirements.txt
pandas 
numpy 
seaborn
matplotlib
pymongo
python-dotenv
certifi
pymongo[srv]==3.6
# -e .


